Allow for 10 tokens per arena
As tokens may be broken during the matches, it makes sense to make spares. 

If we just remade tokens with the same markers on, we risk accidentally putting the same token in twice as a replacement.
The solution to this is to have more than 5 marker IDs allocated to a zone, (in this case 10) and just swap one marker for another if one is damaged.